# Optimize for topology term